// frontend-vite/src/types/generated/LibraryCursor.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LibraryPackMetadata } from "./LibraryPackMetadata";

export type LibraryCursor = { id: string, name: string, file_path: string, click_point_x: number, click_point_y: number, created_at: string, is_pack: boolean, pack_metadata: LibraryPackMetadata | null, };